Moved out monster enum functions
