+++ src/Engine/Objects/SpriteObject.cpp
@@ -16,6 +16,7 @@
 
 #include "Engine/Objects/Actor.h"
 #include "Engine/Objects/ObjectList.h"
+#include "Engine/Objects/MonsterEnumFunctions.h"
 
 #include "Engine/Tables/ItemTable.h"
